# swagger.yaml
openapi: 3.0.0
info:
  title: WhSmith
  contact: {}
  version: '1.0'
servers:
- url: https://avnology.org
  variables: {}
- url: http://localhost:1812
  variables: {}
paths:
...
  /category/add:
    post:
      tags:
      - Category
      summary: Add Category
      operationId: AddCategory
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AddCategoryRequest'
              - example:
                  name:
                    ar: سته
                    en: Sex
            example:
              name:
                ar: سته
                en: Sex
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /category/all:
    get:
      tags:
      - Category
      summary: Get All Category
      operationId: GetAllCategory
      parameters:
      - name: limit
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 10
      - name: category
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: ست
      - name: Accept-Language
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: ar
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /category/one/{id}:
    get:
      tags:
      - Category
      summary: Get Specific Category
      operationId: GetSpecificCategory
      parameters:
      - name: Accept-Language
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: ar
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /category/update/{id}:
    patch:
      tags:
      - Category
      summary: Update Category
      operationId: UpdateCategory
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateCategoryRequest'
              - example:
                  name:
                    ar: سبعه
                    en: Seven
            example:
              name:
                ar: سبعه
                en: Seven
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /cart/all:
    get:
      tags:
      - Cart
      summary: Get All Carts
      operationId: GetAllCarts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /cart/remove/{itemID}:
    delete:
      tags:
      - Cart
      summary: Remove Cart
      operationId: RemoveCart
      parameters:
      - name: itemID
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /cart/update/{itemID}:
    patch:
      tags:
      - Cart
      summary: Update Cart Item Quantity
      operationId: UpdateCartItemQuantity
      parameters:
      - name: itemID
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateCartItemQuantityRequest'
              - example:
                  quantity: 40
            example:
              quantity: 40
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    AddCategoryRequest:
      title: AddCategoryRequest
      required:
      - name
      type: object
      properties:
        name:
          allOf:
          - $ref: '#/components/schemas/Name'
          - {}
      example:
        name:
          ar: سته
          en: Sex
    Name:
      title: Name
      required:
      - ar
      - en
      type: object
      properties:
        ar:
          type: string
        en:
          type: string
      example:
        ar: سته
        en: Sex
    UpdateCategoryRequest:
      title: UpdateCategoryRequest
      required:
      - name
      type: object
      properties:
        name:
          allOf:
          - $ref: '#/components/schemas/Name'
          - {}
      example:
        name:
          ar: سبعه
          en: Seven
...
    UpdateCartItemQuantityRequest:
      title: UpdateCartItemQuantityRequest
      required:
      - quantity
      type: object
      properties:
        quantity:
          type: integer
          format: int32
      example:
        quantity: 40
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...